+++ OpenGl/Assests/Src/render/RenderPass.cpp
@@ -19,7 +19,7 @@ void RenderPass::showUI()
 {
     if (ImGui::TreeNode(getName().c_str()))
     {
-        ImVec2 size(GodClass::getInstance().getWidth(), GodClass::getInstance().getHeight());
+        ImVec2 size(static_cast<float>(GodClass::getInstance().getWidth()), static_cast<float>(GodClass::getInstance().getHeight()));
         float scale = 512.0f / std::max(size.x, size.y);
         size.x *= scale;
         size.y *= scale;
